Main cabin finished.
We could squeeze out more space by using the "wells"
of the rectangular technic frames on the sides, but
it would be a lot of work and I don't know that it
would buy us a lot.

In a real ship those cavities would be filled with
propellant, electronics and stores, leaving the
cabin clean, as we have done.